Refactor: common 모듈 삭제, application 모듈에서 컨버터해서 domain 모듈 메소드 호출, 유저 프로필 수정 시 이름 null 에러 수정
