+++ closet-api/build.gradle
@@ -1,7 +1,6 @@
 project(':closet-api') {
 
     dependencies {
-        implementation project(':closet-common')
         implementation project(':closet-application')
         
         implementation 'org.springframework.boot:spring-boot-starter-web'
