+++ closet-application/build.gradle
@@ -8,7 +8,6 @@ project(':closet-application') {
 
     dependencies {
         implementation project(':closet-domain')
-        implementation project(':closet-common')
         implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
         implementation 'org.springframework.boot:spring-boot-starter-validation'
         implementation 'org.springframework.boot:spring-boot-starter-web'
