+++ closet-application/src/main/java/project/closet/service/dto/response/UserDtoCursorResponse.java
@@ -2,7 +2,7 @@ package project.closet.service.dto.response;
 
 import java.util.List;
 import java.util.UUID;
-import project.closet.SortDirection;
+import project.closet.service.common.SortDirection;
 
 public record UserDtoCursorResponse(
     List<UserDto> data,
