--- closet-common/build.gradle
@@ -1,8 +1,0 @@
-project(':closet-common') {
-    tasks.bootJar {
-        enabled = false
-    }
-    tasks.jar {
-        enabled = true
-    }
-}
