+++ closet-domain/build.gradle
@@ -7,7 +7,6 @@ project(':closet-domain') {
     }
 
     dependencies {
-        implementation project(':closet-common')
         implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
         testImplementation 'com.h2database:h2'
         runtimeOnly 'org.postgresql:postgresql'
