+++ closet-domain/src/main/java/project/closet/feed/repository/impl/FeedRepositoryImpl.java
@@ -8,7 +8,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
-import project.closet.SortDirection;
+import org.hibernate.query.SortDirection;
 import project.closet.feed.entity.Feed;
 import project.closet.feed.repository.FeedRepositoryCustom;
 import project.closet.weather.entity.PrecipitationType;
