+++ closet-domain/src/main/java/project/closet/user/repository/UserRepositoryCustom.java
@@ -2,7 +2,7 @@ package project.closet.user.repository;
 
 import java.util.List;
 import java.util.UUID;
-import project.closet.SortDirection;
+import org.hibernate.query.SortDirection;
 import project.closet.user.entity.Role;
 import project.closet.user.entity.User;
