+++ closet-domain/src/main/java/project/closet/user/repository/impl/UserRepositoryImpl.java
@@ -7,8 +7,8 @@ import java.util.List;
 import java.util.Map;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
+import org.hibernate.query.SortDirection;
 import org.springframework.util.StringUtils;
-import project.closet.SortDirection;
 import project.closet.user.entity.Role;
 import project.closet.user.entity.User;
 import project.closet.user.repository.UserRepositoryCustom;
