+++ settings.gradle
@@ -2,5 +2,4 @@ rootProject.name = 'closet'
 
 include 'closet-api'
 include 'closet-domain'
-include 'closet-common'
 include 'closet-application'
